metadata/requirements.txt
kafka-python
google-api-python-client
stringcase